HikiVisionConnection/requirements.txt
aio-pika==9.5.7
aiormq==6.9.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.3
pamqp==3.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
